Module azure_core::date::timestamp::milliseconds

Treat an OffsetDateTime as a Unix timestamp with milliseconds for the purposes of serde.

Use this module in combination with serde’s #[with] attribute.

When deserializing, the offset is assumed to be UTC.

Modules

option

Treat an Option<OffsetDateTime> as a Unix timestamp with milliseconds for the purposes of serde.

Functions

deserialize

Deserialize an OffsetDateTime from its Unix timestamp with milliseconds

serialize

Serialize an OffsetDateTime as its Unix timestamp with milliseconds