Module azure_core::date::timestamp::milliseconds::option

Treat an Option<OffsetDateTime> as a Unix timestamp with milliseconds for the purposes of serde.

Use this module in combination with serde’s #[with] attribute.

When deserializing, the offset is assumed to be UTC.

Functions

deserialize

Deserialize an Option<OffsetDateTime> from its Unix timestamp with milliseconds

serialize

Serialize an Option<OffsetDateTime> as its Unix timestamp with milliseconds