Enum azure_core::StatusCode

#[repr(u16)]
pub enum StatusCode {

    Continue = 100,
    SwitchingProtocols = 101,
    EarlyHints = 103,
    Ok = 200,
    Created = 201,
    Accepted = 202,
    NonAuthoritativeInformation = 203,
    NoContent = 204,
    ResetContent = 205,
    PartialContent = 206,
    MultiStatus = 207,
    ImUsed = 226,
    MultipleChoice = 300,
    MovedPermanently = 301,
    Found = 302,
    SeeOther = 303,
    NotModified = 304,
    TemporaryRedirect = 307,
    PermanentRedirect = 308,
    BadRequest = 400,
    Unauthorized = 401,
    PaymentRequired = 402,
    Forbidden = 403,
    NotFound = 404,
    MethodNotAllowed = 405,
    NotAcceptable = 406,
    ProxyAuthenticationRequired = 407,
    RequestTimeout = 408,
    Conflict = 409,
    Gone = 410,
    LengthRequired = 411,
    PreconditionFailed = 412,
    PayloadTooLarge = 413,
    UriTooLong = 414,
    UnsupportedMediaType = 415,
    RequestedRangeNotSatisfiable = 416,
    ExpectationFailed = 417,
    ImATeapot = 418,
    MisdirectedRequest = 421,
    UnprocessableEntity = 422,
    Locked = 423,
    FailedDependency = 424,
    TooEarly = 425,
    UpgradeRequired = 426,
    PreconditionRequired = 428,
    TooManyRequests = 429,
    RequestHeaderFieldsTooLarge = 431,
    UnavailableForLegalReasons = 451,
    InternalServerError = 500,
    NotImplemented = 501,
    BadGateway = 502,
    ServiceUnavailable = 503,
    GatewayTimeout = 504,
    HttpVersionNotSupported = 505,
    VariantAlsoNegotiates = 506,
    InsufficientStorage = 507,
    LoopDetected = 508,
    NotExtended = 510,
    NetworkAuthenticationRequired = 511,

}

HTTP response status codes.

As defined by rfc7231 section 6. Read more

Variants

Continue = 100

100 Continue

This interim response indicates that everything so far is OK and that the client should continue the request, or ignore the response if the request is already finished.

SwitchingProtocols = 101

101 Switching Protocols

This code is sent in response to an Upgrade request header from the client, and indicates the protocol the server is switching to.

EarlyHints = 103

103 Early Hints

This status code is primarily intended to be used with the Link header, letting the user agent start preloading resources while the server prepares a response.

Ok = 200

200 Ok

The request has succeeded

Created = 201

201 Created

The request has succeeded and a new resource has been created as a result. This is typically the response sent after POST requests, or some PUT requests.

Accepted = 202

202 Accepted

The request has been received but not yet acted upon. It is noncommittal, since there is no way in HTTP to later send an asynchronous response indicating the outcome of the request. It is intended for cases where another process or server handles the request, or for batch processing.

NonAuthoritativeInformation = 203

203 Non Authoritative Information

This response code means the returned meta-information is not exactly the same as is available from the origin server, but is collected from a local or a third-party copy. This is mostly used for mirrors or backups of another resource. Except for that specific case, the “200 OK” response is preferred to this status.

NoContent = 204

204 No Content

There is no content to send for this request, but the headers may be useful. The user-agent may update its cached headers for this resource with the new ones.

ResetContent = 205

205 Reset Content

Tells the user-agent to reset the document which sent this request.

PartialContent = 206

206 Partial Content

This response code is used when the Range header is sent from the client to request only part of a resource.

MultiStatus = 207

207 Multi-Status

A Multi-Status response conveys information about multiple resources in situations where multiple status codes might be appropriate.

ImUsed = 226

226 Im Used

The server has fulfilled a GET request for the resource, and the response is a representation of the result of one or more instance-manipulations applied to the current instance.

MultipleChoice = 300

300 Multiple Choice

The request has more than one possible response. The user-agent or user should choose one of them. (There is no standardized way of choosing one of the responses, but HTML links to the possibilities are recommended so the user can pick.)

MovedPermanently = 301

301 Moved Permanently

The URL of the requested resource has been changed permanently. The new URL is given in the response.

Found = 302

302 Found

This response code means that the URI of requested resource has been changed temporarily. Further changes in the URI might be made in the future. Therefore, this same URI should be used by the client in future requests.

SeeOther = 303

303 See Other

The server sent this response to direct the client to get the requested resource at another URI with a GET request.

NotModified = 304

304 Not Modified

This is used for caching purposes. It tells the client that the response has not been modified, so the client can continue to use the same cached version of the response.

TemporaryRedirect = 307

307 Temporary Redirect

The server sends this response to direct the client to get the requested resource at another URI with same method that was used in the prior request. This has the same semantics as the 302 Found HTTP response code, with the exception that the user agent must not change the HTTP method used: If a POST was used in the first request, a POST must be used in the second request.

PermanentRedirect = 308

308 Permanent Redirect

This means that the resource is now permanently located at another URI, specified by the Location: HTTP Response header. This has the same semantics as the 301 Moved Permanently HTTP response code, with the exception that the user agent must not change the HTTP method used: If a POST was used in the first request, a POST must be used in the second request.

BadRequest = 400

400 Bad Request

The server could not understand the request due to invalid syntax.

Unauthorized = 401

401 Unauthorized

Although the HTTP standard specifies “unauthorized”, semantically this response means “unauthenticated”. That is, the client must authenticate itself to get the requested response.

PaymentRequired = 402

402 Payment Required

This response code is reserved for future use. The initial aim for creating this code was using it for digital payment systems, however this status code is used very rarely and no standard convention exists.

Forbidden = 403

403 Forbidden

The client does not have access rights to the content; that is, it is unauthorized, so the server is refusing to give the requested resource. Unlike 401, the client’s identity is known to the server.

NotFound = 404

404 Not Found

The server can not find requested resource. In the browser, this means the URL is not recognized. In an API, this can also mean that the endpoint is valid but the resource itself does not exist. Servers may also send this response instead of 403 to hide the existence of a resource from an unauthorized client. This response code is probably the most famous one due to its frequent occurrence on the web.

MethodNotAllowed = 405

405 Method Not Allowed

The request method is known by the server but has been disabled and cannot be used. For example, an API may forbid DELETE-ing a resource. The two mandatory methods, GET and HEAD, must never be disabled and should not return this error code.

NotAcceptable = 406

406 Not Acceptable

This response is sent when the web server, after performing server-driven content negotiation, doesn’t find any content that conforms to the criteria given by the user agent.

ProxyAuthenticationRequired = 407

407 Proxy Authentication Required

This is similar to 401 but authentication is needed to be done by a proxy.

RequestTimeout = 408

408 Request Timeout

This response is sent on an idle connection by some servers, even without any previous request by the client. It means that the server would like to shut down this unused connection. This response is used much more since some browsers, like Chrome, Firefox 27+, or IE9, use HTTP pre-connection mechanisms to speed up surfing. Also note that some servers merely shut down the connection without sending this message.

Conflict = 409

409 Conflict

This response is sent when a request conflicts with the current state of the server.

Gone = 410

410 Gone

This response is sent when the requested content has been permanently deleted from server, with no forwarding address. Clients are expected to remove their caches and links to the resource. The HTTP specification intends this status code to be used for “limited-time, promotional services”. APIs should not feel compelled to indicate resources that have been deleted with this status code.

LengthRequired = 411

411 Length Required

Server rejected the request because the Content-Length header field is not defined and the server requires it.

PreconditionFailed = 412

412 Precondition Failed

The client has indicated preconditions in its headers which the server does not meet.

PayloadTooLarge = 413

413 Payload Too Large

Request entity is larger than limits defined by server; the server might close the connection or return an Retry-After header field.

UriTooLong = 414

414 URI Too Long

The URI requested by the client is longer than the server is willing to interpret.

UnsupportedMediaType = 415

415 Unsupported Media Type

The media format of the requested data is not supported by the server, so the server is rejecting the request.

RequestedRangeNotSatisfiable = 416

416 Requested Range Not Satisfiable

The range specified by the Range header field in the request can’t be fulfilled; it’s possible that the range is outside the size of the target URI’s data.

ExpectationFailed = 417

417 Expectation Failed

This response code means the expectation indicated by the Expect request header field can’t be met by the server.

ImATeapot = 418

418 I’m a teapot

The server refuses the attempt to brew coffee with a teapot.

MisdirectedRequest = 421

421 Misdirected Request

The request was directed at a server that is not able to produce a response. This can be sent by a server that is not configured to produce responses for the combination of scheme and authority that are included in the request URI.

UnprocessableEntity = 422

422 Unprocessable Entity

The request was well-formed but was unable to be followed due to semantic errors.

Locked = 423

423 Locked

The resource that is being accessed is locked.

FailedDependency = 424

424 Failed Dependency

The request failed because it depended on another request and that request failed (e.g., a PROPPATCH).

TooEarly = 425

425 Too Early

Indicates that the server is unwilling to risk processing a request that might be replayed.

UpgradeRequired = 426

426 Upgrade Required

The server refuses to perform the request using the current protocol but might be willing to do so after the client upgrades to a different protocol. The server sends an Upgrade header in a 426 response to indicate the required protocol(s).

PreconditionRequired = 428

428 Precondition Required

The origin server requires the request to be conditional. This response is intended to prevent the ‘lost update’ problem, where a client GETs a resource’s state, modifies it, and PUTs it back to the server, when meanwhile a third party has modified the state on the server, leading to a conflict.

TooManyRequests = 429

429 Too Many Requests

The user has sent too many requests in a given amount of time (“rate limiting”).

RequestHeaderFieldsTooLarge = 431

431 Request Header Fields Too Large

The server is unwilling to process the request because its header fields are too large. The request may be resubmitted after reducing the size of the request header fields.

UnavailableForLegalReasons = 451

451 Unavailable For Legal Reasons

The user-agent requested a resource that cannot legally be provided, such as a web page censored by a government.

InternalServerError = 500

500 Internal Server Error

The server has encountered a situation it doesn’t know how to handle.

NotImplemented = 501

501 Not Implemented

The request method is not supported by the server and cannot be handled. The only methods that servers are required to support (and therefore that must not return this code) are GET and HEAD.

BadGateway = 502

502 Bad Gateway

This error response means that the server, while working as a gateway to get a response needed to handle the request, got an invalid response.

ServiceUnavailable = 503

503 Service Unavailable

The server is not ready to handle the request. Common causes are a server that is down for maintenance or that is overloaded. Note that together with this response, a user-friendly page explaining the problem should be sent. This responses should be used for temporary conditions and the Retry-After: HTTP header should, if possible, contain the estimated time before the recovery of the service. The webmaster must also take care about the caching-related headers that are sent along with this response, as these temporary condition responses should usually not be cached.

GatewayTimeout = 504

504 Gateway Timeout

This error response is given when the server is acting as a gateway and cannot get a response in time.

HttpVersionNotSupported = 505

505 HTTP Version Not Supported

The HTTP version used in the request is not supported by the server.

VariantAlsoNegotiates = 506

506 Variant Also Negotiates

The server has an internal configuration error: the chosen variant resource is configured to engage in transparent content negotiation itself, and is therefore not a proper end point in the negotiation process.

InsufficientStorage = 507

507 Insufficient Storage

The server is unable to store the representation needed to complete the request.

LoopDetected = 508

508 Loop Detected

The server detected an infinite loop while processing the request.

NotExtended = 510

510 Not Extended

Further extensions to the request are required for the server to fulfil it.

NetworkAuthenticationRequired = 511

511 Network Authentication Required

The 511 status code indicates that the client needs to authenticate to gain network access.

Implementations

impl StatusCode

pub fn is_informational(&self) -> bool

Returns true if the status code is 1xx range.

If this returns true it indicates that the request was received, continuing process.

pub fn is_success(&self) -> bool

Returns true if the status code is the 2xx range.

If this returns true it indicates that the request was successfully received, understood, and accepted.

pub fn is_redirection(&self) -> bool

Returns true if the status code is the 3xx range.

If this returns true it indicates that further action needs to be taken in order to complete the request.

pub fn is_client_error(&self) -> bool

Returns true if the status code is the 4xx range.

If this returns true it indicates that the request contains bad syntax or cannot be fulfilled.

pub fn is_server_error(&self) -> bool

Returns true if the status code is the 5xx range.

If this returns true it indicates that the server failed to fulfill an apparently valid request.

pub fn canonical_reason(&self) -> &'static str

The canonical reason for a given status code