azure_core::headers
Trait Header

pub trait Header {
    // Required methods
    fn name(&self) -> HeaderName;
    fn value(&self) -> HeaderValue;
}

View a type as an HTTP header.

Ad interim there are two default functions: add_to_builder and add_to_request.

While not restricted by the type system, please add HTTP headers only. In particular, do not interact with the body of the request.

As soon as the migration to the pipeline architecture will be complete we will phase out add_to_builder.

Required Methods

fn name(&self) -> HeaderName

fn value(&self) -> HeaderValue