Module azure_core::xml

Modules

text_content

Provides helper functions to serialization and deserialization of types (usually enums) as a text content of an element and intended to use with #[serde(with = "...")], #[serde(deserialize_with = "...")] and #[serde(serialize_with = "...")].

Functions

read_xml

Reads the XML from bytes.

read_xml_str

Reads the XML from bytes.

to_xml

to_xml_with_root