Crate azure_core

azure_core

Microsoft is developing the official Azure SDK for Rust crates and has no plans to update this unofficial crate. In the future we may release an official version that may have a different package name. If releasing an official version of this crate is important to you let us know.

Source for this crate can now be found in https://github.com/Azure/azure-sdk-for-rust/tree/legacy. To monitor for an official, supported version of this crate, see https://aka.ms/azsdk/releases.

Core types and traits for the Rust Azure SDK.

This crate is part of the unofficial Azure SDK effort in Rust. For more information on the project and an overview of other crates, please refer to our GitHub repository.

It is a library that provides cross-cutting services to other client libraries. Please see the general guidelines.

License: MIT

Re-exports

pub use error::Error;

pub use error::Result;

pub use sleep::sleep;

Modules

auth

Azure authentication and authorization.

authority_hosts

A list of known Azure authority hosts

base64

content_type

Constants related to the Content-Type header

date

Azure date and time parsing and formatting

error

headers

Azure HTTP headers.

hmac

lro

parsing

Parser helper utilities.

prelude

The Azure Core prelude.

query_param

Constants related to query parameters

request_options

resource_manager_endpoint

Endpoints for Azure Resource Manager in different Azure clouds

sleep

tokio

util

An assortment of helper utilities.

xml

Macros

create_enum

The following macro invocation:

ensure

Return early with an error if a condition is not satisfied.

ensure_eq

Return early with an error if two expressions are not equal to each other.

ensure_ne

Return early with an error if two expressions are equal to each other.

format_err

A convenient way to create a new error using the normal formatting infrastructure

future

Declare a Future with the given name

operation

Helper for constructing operations

request_header

The following macro invocation:

request_option

The following macro invocation:

request_query

The following macro invocation:

setters

Creates setter methods

static_url

Structs

BytesStream

Convenience struct that maps a bytes::Bytes buffer into a stream.

ClientOptions

Client options allow customization of policies, retry options, and more.

CollectedResponse

A response with the body collected as bytes

Context

Pipeline execution context.

CustomHeaders

CustomHeadersPolicy

Etag

ExponentialRetryOptions

Options for how an exponential retry strategy should behave.

ExponentialRetryPolicy

Retry policy with exponential back-off.

FixedRetryOptions

Options for how a fixed retry strategy should behave.

FixedRetryPolicy

Retry policy with a fixed back-off.

NoRetryPolicy

Retry policy that does not retry.

Pageable

A pageable stream that yields items of type T

Pipeline

Execution pipeline.

Request

A pipeline request.

Response

An HTTP Response.

ResponseBody

A response body stream

RetryOptions

Specify how retries should behave.

TelemetryOptions

Telemetry options.

TelemetryPolicy

TimeoutPolicy

TransportOptions

Transport options.

TransportPolicy

Url

A parsed URL record.

Enums

Body

An HTTP Body.

LeaseAction

LeaseDuration

LeaseState

LeaseStatus

Method

HTTP request methods.

StatusCode

HTTP response status codes.

Constants

EMPTY_BODY

An empty HTTP body.

Traits

AppendToUrlQuery

Add a new query pair into the target URL’s query string.

Continuable

A type that can yield an optional continuation token

Header

View a type as an HTTP header.

HttpClient

An HTTP client which can send requests.

Policy

A pipeline policy.

RetryPolicy

A retry policy.

SeekableStream

Enable a type implementing AsyncRead to be consumed as if it were a Stream of Bytes.

Functions

from_json

Reads the XML from bytes.

new_http_client

Construct a new HttpClient

to_json

Serialize a type to json.

Type Aliases

PolicyResult

A specialized Result type for policies.

RequestId

A unique identifier for a request.

SessionToken

A unique session token.