azure_core
Module request_options

Structs

Accept

AcceptEncoding

Advertises which content encoding the client is able to understand.

ActivityId

App

ClientRequestId

ClientVersion

ContentDisposition

ContentEncoding

ContentLanguage

ContentLength

ContentRange

ContentType

The Content Type indicates the media type of the request body

Continuation

Delimiter

Set delimiter for the request

IfModifiedSince

IfTags

LeaseBreakPeriod

LeaseId

MaxItemCount

The max number of items in the collection

MaxResults

Metadata

NextMarker

Prefix

ProposedLeaseId

SequenceNumber

SourceLeaseId

Timeout

User

The (friendly) name of the user making the request

UserAgent

Version

Enums

IfMatchCondition

IfModifiedSinceCondition

IfSequenceNumber

Conditional request header based on the value of the object’s sequence number

IfSourceMatchCondition

IfSourceModifiedSinceCondition

LeaseDuration

Range